Units as free functions (#80)
* all tests passing

* Units changed to utils

* Changed to explicit imports of utils.

* Removed unused variables.

* units.rst changed to utils.rst

* Units changed to Utils

* excludes UtilError
