--- doc/source/api/units.rst
@@ -1,11 +1,0 @@
-.. _ref_units:
-
-=====
-Units
-=====
-
-.. automodule:: ansys.units.units
-   :members:
-   :show-inheritance:
-   :undoc-members:
-   :exclude-members: __weakref__, __dict__, __init__, parse_temperature_units
