+++ doc/source/api/utils.rst
@@ -1,0 +1,11 @@
+.. _ref_units:
+
+=====
+Utils
+=====
+
+.. automodule:: ansys.units.utils
+   :members:
+   :show-inheritance:
+   :undoc-members:
+   :exclude-members: __weakref__, __dict__, __init__, parse_temperature_units, UtilError
