+++ src/ansys/units/systems.py
@@ -1,5 +1,6 @@
 """Provides the ``UnitSystem`` class."""
 import ansys.units as ansunits
+from ansys.units.utils import si_data
 
 
 class UnitSystem:
@@ -27,8 +28,6 @@ class UnitSystem:
     """
 
     def __init__(self, name: str = None, base_units: list = None, unit_sys: str = None):
-        self._units = ansunits.Units()
-
         if name and unit_sys or base_units and unit_sys:
             raise UnitSystemError.EXCESSIVE_PARAMETERS()
 
@@ -80,7 +79,7 @@ class UnitSystem:
             dimensions=quantity.dimensions, unit_sys=self._base_units
         )
 
-        _, si_multiplier, si_offset = self._units.si_data(new_dim.units)
+        _, si_multiplier, si_offset = si_data(new_dim.units)
         new_value = (quantity.si_value / si_multiplier) - si_offset
 
         return ansunits.Quantity(value=new_value, units=new_dim.units)
